src/pipelines/payment_processing/bigtablequerytesting/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
google-cloud-bigtable==2.27.0
